+++ robots/src/main/java/visill/robot/App.java
@@ -1,0 +1,30 @@
+package visill.robot;
+
+import visill.robot.gui.MainApplicationFrame;
+
+import java.awt.Frame;
+
+import javax.swing.SwingUtilities;
+import javax.swing.UIManager;
+
+public class App {
+    public String getGreeting() {
+        return "Hello World!";
+    }
+    public static void main(String[] args) {
+        try {
+            UIManager.setLookAndFeel("javax.swing.plaf.nimbus.NimbusLookAndFeel");
+//        UIManager.setLookAndFeel("javax.swing.plaf.metal.MetalLookAndFeel");
+//        UIManager.setLookAndFeel(UIManager.getSystemLookAndFeelClassName());
+//        UIManager.setLookAndFeel(UIManager.getCrossPlatformLookAndFeelClassName());
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+        SwingUtilities.invokeLater(() -> {
+            MainApplicationFrame frame = new MainApplicationFrame();
+            frame.pack();
+            frame.setVisible(true);
+            frame.setExtendedState(Frame.MAXIMIZED_BOTH);
+        });
+    }
+}
