+++ robots/src/main/java/visill/robot/gui/LogWindow.java
@@ -1,4 +1,4 @@
-package gui;
+package visill.robot.gui;
 
 import java.awt.BorderLayout;
 import java.awt.EventQueue;
@@ -7,9 +7,9 @@ import java.awt.TextArea;
 import javax.swing.JInternalFrame;
 import javax.swing.JPanel;
 
-import log.LogChangeListener;
-import log.LogEntry;
-import log.LogWindowSource;
+import visill.robot.log.LogChangeListener;
+import visill.robot.log.LogEntry;
+import visill.robot.log.LogWindowSource;
 
 public class LogWindow extends JInternalFrame implements LogChangeListener
 {
