+++ robots/src/main/java/visill/robot/log/LogChangeListener.java
@@ -1,4 +1,4 @@
-package log;
+package visill.robot.log;
 
 public interface LogChangeListener
 {
