+++ robots/src/main/java/visill/robot/log/LogEntry.java
@@ -1,4 +1,4 @@
-package log;
+package visill.robot.log;
 
 public class LogEntry
 {
