+++ robots/src/main/java/visill/robot/log/LogLevel.java
@@ -1,4 +1,4 @@
-package log;
+package visill.robot.log;
 
 public enum LogLevel
 {
