+++ robots/src/main/java/visill/robot/log/LogWindowSource.java
@@ -1,4 +1,4 @@
-package log;
+package visill.robot.log;
 
 import java.util.ArrayList;
 import java.util.Collections;
