+++ robots/src/main/java/visill/robot/log/Logger.java
@@ -1,4 +1,4 @@
-package log;
+package visill.robot.log;
 
 public final class Logger
 {
