+++ robots/src/test/java/visill/robot/AppTest.java
@@ -1,7 +1,7 @@
 /*
  * This source file was generated by the Gradle 'init' task
  */
-package org.example;
+package visill.robot;
 
 import org.junit.Test;
 import static org.junit.Assert.*;
